refactor: :art: reorganize project structure in a gptcli folder
If we want to be able to add a setup file that contains a script that allows to execute the program using 'gptcli' rather than 'python3 main.py', we need to have this project structure. It also distinguished between 'main.py' and the folder named 'main'.

We might have import errors