feat(helper): Add error information retrieved from server to logs.
An example of this occuring would be when the user's API key does not grant them access to one of the model options of gptcli.

It is unknown if we will always receive an error json when there is an error during a POST. So, an error could be that 'error' does not exist as a key.