refactor(project): Remove unused profiles file.
The Profiles feature would be functionally the same as adding an external file for context via the --filepath parameter. They would be packaged up into the same message when calling to the API. Not to mention that OpenAI constantly makes soft changes to their models that could break the hardcoded profiles in GPTCLI, there is also a lack of a good source for ChatGPT profiles, therefore requiring constant  updating and constant risk of breaking. As such, it should be removed.