feat(message): Expand Message functionality.
Add tests to cover the functionality we have covered in previous commits. Allow the Messages object to return the number of messages using len(). Allow creation of a Messages object with the messages parameter in __init__.  Fix the docstrings and logs in the messages.py to have periods (Logs and Docs should have proper punctuation and other corrections).