feat(project): :construction: Add foundation for Single-Exchange mode.
This mode is used when we only want to ask ChatGPT 1 thing and then exit, without needing to enter a chat. This makes GPTCLI useful for execution by other programs through the terminal. It also allows the possibility of multithreading sometime down the line. This commit also includes improvements to definitions.py and some changes as to how GPTCLI will get the API key for it's chats or se's.