refactor(chat): Improve chat readability, coverage, reply process, and abort process.
A lot of changes here. For the tests, we have finally included some
tests to improve coverage for a core feature of the project. We have
removed the sseclient and decided to mimic streaming manually.
This gives a smoother stream removes a useless dependency.

This commit is unstable. Some tests fail by design. Do not point HEAD to
this commit in production.