refactor(project): Major changes going in.
Resolved a bug related to streaming taking too long (This was due to a misunderstanding with how the Session class from Requests works). The solution is, print the output as it arrives in openai_api_helper module and return the full result as a Message or Messages object once everything is printed. However, this breaks another feature; our 'output' flag. This is beause the flag requires a Response object to be sent to it for parsing and not a Message(s) object. To resolve this we create a class in openai_api_helper called SingleExchange. The idea is we lock 'output' to only the Single Exchange method of communication, and return only Response objects without printing the content yet. Another major feature is making the project compile to C code using mypycify (this should make the interface feel more snappy due to increased execution speed). Also, improve code coverage of chat.py. Fix some tests to reflect the new features and refactors. Add o1 as option to be turned on for when/if OpenAI gives access. Improve import structure. Add logger where loggers need to be added. Improve some naming of variables.