build: Make build of project with C code optional.
The whole project as of this commit is ready to be converted to C code using mypycify. However, doing this will generate a platform specific wheel file. This is not suitable for pypi where the target audience is much wider than my local machine. As such, comment it out until a better solution can be found that will make platform specific wheel files and non available to the any repo user.