fix(build): Issue where Macs and WSL systems do not have gptcli in PATH.
This issue did not occur with Ubuntu because it was perfectly capable of using setup.py to build the wheel. The Mac and WSL systems do not do this by default. Therefore, we must commit fully to using pyproject.toml to avoid cross system build inconsistencies, and fix a mistake with in pyproject.toml where the program in the PATH was 'scripts' and not 'gptcli'.

BREAKING CHANGE: We are removing the feature to build the project with mypycify because it requires setup.py to work. And I am unwilling to find a workaround.